refactor(ask): move spinner into ask
